requirements/dev.txt
-r base.txt

ruff==0.8.4
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
pre-commit==4.0.1
